app/api/requirements.txt
boto3
flask
flask_cors
gunicorn
pillow
dotenv